treewide: Refactor SensorSystemDTO into its own file
