+++ lib/models/sensor_system/sensor_system_dto.dart
