+++ test/models/sensor_system/sensor_system_test.dart
@@ -1,7 +1,7 @@
 import 'package:environment_monitor/consts/sensor_location_type.dart';
 import 'package:environment_monitor/consts/sensor_status.dart';
 import 'package:environment_monitor/models/sensor_location/sensor_location.dart';
-import 'package:environment_monitor/models/sensor_system/sensor_system.dart';
+import 'package:environment_monitor/models/sensor_system/sensor_system_dto.dart';
 import 'package:environment_monitor/models/user/unconv_user.dart';
 import 'package:flutter_test/flutter_test.dart';
